requirements.txt
dbt==0.20.0
opentelemetry-sdk==1.3.0
opentelemetry-exporter-otlp==1.3.0
